requirements.txt
Django
django-mssql-backend
pyodbc
pillow
opencv-python
numpy
asgiref
sqlparse
tzdata
mediapipe
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
google-auth
comtypes
pdf2image
